+++ src/game/stateless_systems/gun_system.cpp
@@ -1,3 +1,4 @@
+#include "augs/ensure_rel.h"
 #include "augs/misc/randomization.h"
 
 #include "game/cosmos/cosmos.h"
@@ -735,7 +736,7 @@ void gun_system::launch_shots_due_to_pressed_triggers(const logic_step step) {
 									}
 								}
 
-								ensure_greater(bullet_stacks.size(), 0);
+								ensure_greater(static_cast<int>(bullet_stacks.size()), 0);
 
 								for (const auto single_bullet_or_pellet_stack_id : bullet_stacks) {
 									const auto single_bullet_or_pellet_stack = cosm[single_bullet_or_pellet_stack_id];
